refactor: confirm firefly modal
